sprint01/requirements.txt
torch>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
pytest>=7.0.0
# Optional but recommended:
# matplotlib>=3.7.0   # for plots in analysis/visualize.py
# scipy>=1.10.0       # for Wilcoxon test in analysis/statistics.py
